bgfx: add -bgfx_vec_supersample option and fractional render target scales
Replace the hardcoded VEC_SUPERSAMPLE constant with a -bgfx_vec_supersample
option (1-2, default 2) controlling the vector FBO supersampling factor;
1 quarters the FBO pixel count for bandwidth-constrained setups (the
analytic AA in fs_vector_line keeps line quality acceptable).

Allow fractional "scale" values on chain JSON render targets: the target
reader no longer rounds the scale to an integer, and bgfx_target carries it
as float (scaled dimensions are rounded to nearest with a 1px floor, so
existing integer-scale chains are unchanged). The vector chains ladder
bloom accumulators switch from fixed 960x540-based custom targets to
window-proportional native targets (scale 0.5 .. 1/128), which removes the
bloom anisotropy on non-16:9 windows, keeps acc0 matched 1:1 with
bloom_lvl0, and tracks window resizes automatically.

+++ src/osd/modules/render/bgfx/target.h
@@ -28,7 +28,7 @@ class bgfx_target : public bgfx_texture_handle_provider
 {
 public:
 	bgfx_target(std::string name, bgfx::TextureFormat::Enum format, uint16_t width, uint16_t height, uint16_t xprescale, uint16_t yprescale,
-		uint32_t style, bool double_buffer, bool filter, uint16_t scale, uint32_t screen);
+		uint32_t style, bool double_buffer, bool filter, float scale, uint32_t screen);
 	bgfx_target(void *handle, uint16_t width, uint16_t height);
 	virtual ~bgfx_target();
 
@@ -41,7 +41,7 @@ public:
 	bool                        double_buffered() const { return m_double_buffer; }
 	uint32_t                    style() const { return m_style; }
 	bool                        filter() const { return m_filter; }
-	uint16_t                    scale() const { return m_scale; }
+	float                       scale() const { return m_scale; }
 	uint32_t                    screen_index() const { return m_screen; }
 	uint16_t                    raw_width() const { return m_width; }
 	uint16_t                    raw_height() const { return m_height; }
@@ -72,7 +72,7 @@ private:
 	bool                        m_double_buffer;
 	uint32_t                    m_style;
 	bool                        m_filter;
-	uint16_t                    m_scale;
+	float                       m_scale;
 
 	int32_t                     m_screen;
